# jij-service/api/server.gen.go
//go:build go1.22

// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.7.1 DO NOT EDIT.
package api

import (
	"bytes"
	"context"
	"encoding/json"
	"fmt"
	"net/http"

	"github.com/oapi-codegen/runtime"
)

// Leaderboard defines model for Leaderboard.
type Leaderboard struct {
	Entries *[]LeaderboardEntry `json:"entries,omitempty"`
}

// LeaderboardEntry defines model for LeaderboardEntry.
type LeaderboardEntry struct {
	ProfileId *string  `json:"profileId,omitempty"`
	Rank      *float32 `json:"rank,omitempty"`
	Score     *int     `json:"score,omitempty"`
}

// Pong defines model for Pong.
type Pong struct {
	Ping string `json:"ping"`
}

// ProfileComponent defines model for ProfileComponent.
type ProfileComponent struct {
	ComponentId *string               `json:"componentId,omitempty"`
	Data        *ProfileComponentData `json:"data,omitempty"`
}

// ProfileComponentData defines model for ProfileComponentData.
type ProfileComponentData map[string]map[string]interface{}

// ProfileComponentPatch defines model for ProfileComponentPatch.
type ProfileComponentPatch struct {
	Data *ProfileComponentData `json:"data,omitempty"`
}

// ProfileRunDetails defines model for ProfileRunDetails.
type ProfileRunDetails struct {
	AsOf      *string  `json:"asOf,omitempty"`
	Duration  *float32 `json:"duration,omitempty"`
	LevelName *string  `json:"levelName,omitempty"`
	Score     *int     `json:"score,omitempty"`
}

// PatchProfileProfileIdComponentComponentIdJSONRequestBody defines body for PatchProfileProfileIdComponentComponentId for application/json ContentType.
type PatchProfileProfileIdComponentComponentIdJSONRequestBody = ProfileComponentPatch

// PostProfileProfileIdRunRunIdJSONRequestBody defines body for PostProfileProfileIdRunRunId for application/json ContentType.
type PostProfileProfileIdRunRunIdJSONRequestBody = ProfileRunDetails

// ServerInterface represents all server handlers.
type ServerInterface interface {

	// (GET /leaderboard/{leaderboardName})
	GetLeaderboardLeaderboardName(w http.ResponseWriter, r *http.Request, leaderboardName string)

	// (GET /ping)
	GetPing(w http.ResponseWriter, r *http.Request)

	// (GET /profile/{profileId}/component/{componentId})
	GetProfileProfileIdComponentComponentId(w http.ResponseWriter, r *http.Request, profileId string, componentId string)

	// (PATCH /profile/{profileId}/component/{componentId})
	PatchProfileProfileIdComponentComponentId(w http.ResponseWriter, r *http.Request, profileId string, componentId string)

	// (POST /profile/{profileId}/run/{runId})
	PostProfileProfileIdRunRunId(w http.ResponseWriter, r *http.Request, profileId string, runId string)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// GetLeaderboardLeaderboardName operation middleware
func (siw *ServerInterfaceWrapper) GetLeaderboardLeaderboardName(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "leaderboardName" -------------
	var leaderboardName string

	err = runtime.BindStyledParameterWithOptions("simple", "leaderboardName", r.PathValue("leaderboardName"), &leaderboardName, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "leaderboardName", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetLeaderboardLeaderboardName(w, r, leaderboardName)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetPing operation middleware
func (siw *ServerInterfaceWrapper) GetPing(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetPing(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetProfileProfileIdComponentComponentId operation middleware
func (siw *ServerInterfaceWrapper) GetProfileProfileIdComponentComponentId(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "profileId" -------------
	var profileId string

	err = runtime.BindStyledParameterWithOptions("simple", "profileId", r.PathValue("profileId"), &profileId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "profileId", Err: err})
		return
	}

	// ------------- Path parameter "componentId" -------------
	var componentId string

	err = runtime.BindStyledParameterWithOptions("simple", "componentId", r.PathValue("componentId"), &componentId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "componentId", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetProfileProfileIdComponentComponentId(w, r, profileId, componentId)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// PatchProfileProfileIdComponentComponentId operation middleware
func (siw *ServerInterfaceWrapper) PatchProfileProfileIdComponentComponentId(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "profileId" -------------
	var profileId string

	err = runtime.BindStyledParameterWithOptions("simple", "profileId", r.PathValue("profileId"), &profileId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "profileId", Err: err})
		return
	}

	// ------------- Path parameter "componentId" -------------
	var componentId string

	err = runtime.BindStyledParameterWithOptions("simple", "componentId", r.PathValue("componentId"), &componentId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "componentId", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PatchProfileProfileIdComponentComponentId(w, r, profileId, componentId)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// PostProfileProfileIdRunRunId operation middleware
func (siw *ServerInterfaceWrapper) PostProfileProfileIdRunRunId(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "profileId" -------------
	var profileId string

	err = runtime.BindStyledParameterWithOptions("simple", "profileId", r.PathValue("profileId"), &profileId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "profileId", Err: err})
		return
	}

	// ------------- Path parameter "runId" -------------
	var runId string

	err = runtime.BindStyledParameterWithOptions("simple", "runId", r.PathValue("runId"), &runId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "runId", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PostProfileProfileIdRunRunId(w, r, profileId, runId)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{})
}

// ServeMux is an abstraction of [http.ServeMux].
type ServeMux interface {
	HandleFunc(pattern string, handler func(http.ResponseWriter, *http.Request))
	http.Handler
}

type StdHTTPServerOptions struct {
	BaseURL          string
	BaseRouter       ServeMux
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, m ServeMux) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseRouter: m,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, m ServeMux, baseURL string) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseURL:    baseURL,
		BaseRouter: m,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options StdHTTPServerOptions) http.Handler {
	m := options.BaseRouter

	if m == nil {
		m = http.NewServeMux()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/leaderboard/{leaderboardName}", wrapper.GetLeaderboardLeaderboardName)
	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/ping", wrapper.GetPing)
	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/profile/{profileId}/component/{componentId}", wrapper.GetProfileProfileIdComponentComponentId)
	m.HandleFunc(http.MethodPatch+" "+options.BaseURL+"/profile/{profileId}/component/{componentId}", wrapper.PatchProfileProfileIdComponentComponentId)
	m.HandleFunc(http.MethodPost+" "+options.BaseURL+"/profile/{profileId}/run/{runId}", wrapper.PostProfileProfileIdRunRunId)

	return m
}

type GetLeaderboardLeaderboardNameRequestObject struct {
	LeaderboardName string `json:"leaderboardName"`
}

type GetLeaderboardLeaderboardNameResponseObject interface {
	VisitGetLeaderboardLeaderboardNameResponse(w http.ResponseWriter) error
}

type GetLeaderboardLeaderboardName200JSONResponse Leaderboard

func (response GetLeaderboardLeaderboardName200JSONResponse) VisitGetLeaderboardLeaderboardNameResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)
	_, err := buf.WriteTo(w)
	return err
}

type GetPingRequestObject struct {
}

type GetPingResponseObject interface {
	VisitGetPingResponse(w http.ResponseWriter) error
}

type GetPing200JSONResponse Pong

func (response GetPing200JSONResponse) VisitGetPingResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)
	_, err := buf.WriteTo(w)
	return err
}

type GetProfileProfileIdComponentComponentIdRequestObject struct {
	ProfileId   string `json:"profileId"`
	ComponentId string `json:"componentId"`
}

type GetProfileProfileIdComponentComponentIdResponseObject interface {
	VisitGetProfileProfileIdComponentComponentIdResponse(w http.ResponseWriter) error
}

type GetProfileProfileIdComponentComponentId200JSONResponse ProfileComponent

func (response GetProfileProfileIdComponentComponentId200JSONResponse) VisitGetProfileProfileIdComponentComponentIdResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)
	_, err := buf.WriteTo(w)
	return err
}

type PatchProfileProfileIdComponentComponentIdRequestObject struct {
	ProfileId   string `json:"profileId"`
	ComponentId string `json:"componentId"`
	Body        *PatchProfileProfileIdComponentComponentIdJSONRequestBody
}

type PatchProfileProfileIdComponentComponentIdResponseObject interface {
	VisitPatchProfileProfileIdComponentComponentIdResponse(w http.ResponseWriter) error
}

type PatchProfileProfileIdComponentComponentId200JSONResponse ProfileComponent

func (response PatchProfileProfileIdComponentComponentId200JSONResponse) VisitPatchProfileProfileIdComponentComponentIdResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)
	_, err := buf.WriteTo(w)
	return err
}

type PostProfileProfileIdRunRunIdRequestObject struct {
	ProfileId string `json:"profileId"`
	RunId     string `json:"runId"`
	Body      *PostProfileProfileIdRunRunIdJSONRequestBody
}

type PostProfileProfileIdRunRunIdResponseObject interface {
	VisitPostProfileProfileIdRunRunIdResponse(w http.ResponseWriter) error
}

type PostProfileProfileIdRunRunId200Response struct {
}

func (response PostProfileProfileIdRunRunId200Response) VisitPostProfileProfileIdRunRunIdResponse(w http.ResponseWriter) error {
	w.WriteHeader(200)
	return nil
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {

	// (GET /leaderboard/{leaderboardName})
	GetLeaderboardLeaderboardName(ctx context.Context, request GetLeaderboardLeaderboardNameRequestObject) (GetLeaderboardLeaderboardNameResponseObject, error)

	// (GET /ping)
	GetPing(ctx context.Context, request GetPingRequestObject) (GetPingResponseObject, error)

	// (GET /profile/{profileId}/component/{componentId})
	GetProfileProfileIdComponentComponentId(ctx context.Context, request GetProfileProfileIdComponentComponentIdRequestObject) (GetProfileProfileIdComponentComponentIdResponseObject, error)

	// (PATCH /profile/{profileId}/component/{componentId})
	PatchProfileProfileIdComponentComponentId(ctx context.Context, request PatchProfileProfileIdComponentComponentIdRequestObject) (PatchProfileProfileIdComponentComponentIdResponseObject, error)

	// (POST /profile/{profileId}/run/{runId})
	PostProfileProfileIdRunRunId(ctx context.Context, request PostProfileProfileIdRunRunIdRequestObject) (PostProfileProfileIdRunRunIdResponseObject, error)
}

type StrictHandlerFunc func(ctx context.Context, w http.ResponseWriter, r *http.Request, request any) (any, error)
type StrictMiddlewareFunc func(f StrictHandlerFunc, operationID string) StrictHandlerFunc

type StrictHTTPServerOptions struct {
	RequestErrorHandlerFunc  func(w http.ResponseWriter, r *http.Request, err error)
	ResponseErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: StrictHTTPServerOptions{
		RequestErrorHandlerFunc: func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		},
		ResponseErrorHandlerFunc: func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusInternalServerError)
		},
	}}
}

func NewStrictHandlerWithOptions(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc, options StrictHTTPServerOptions) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: options}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
	options     StrictHTTPServerOptions
}

// GetLeaderboardLeaderboardName operation middleware
func (sh *strictHandler) GetLeaderboardLeaderboardName(w http.ResponseWriter, r *http.Request, leaderboardName string) {
	var request GetLeaderboardLeaderboardNameRequestObject

	request.LeaderboardName = leaderboardName

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.GetLeaderboardLeaderboardName(ctx, request.(GetLeaderboardLeaderboardNameRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetLeaderboardLeaderboardName")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(GetLeaderboardLeaderboardNameResponseObject); ok {
		if err := validResponse.VisitGetLeaderboardLeaderboardNameResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// GetPing operation middleware
func (sh *strictHandler) GetPing(w http.ResponseWriter, r *http.Request) {
	var request GetPingRequestObject

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.GetPing(ctx, request.(GetPingRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetPing")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(GetPingResponseObject); ok {
		if err := validResponse.VisitGetPingResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// GetProfileProfileIdComponentComponentId operation middleware
func (sh *strictHandler) GetProfileProfileIdComponentComponentId(w http.ResponseWriter, r *http.Request, profileId string, componentId string) {
	var request GetProfileProfileIdComponentComponentIdRequestObject

	request.ProfileId = profileId
	request.ComponentId = componentId

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.GetProfileProfileIdComponentComponentId(ctx, request.(GetProfileProfileIdComponentComponentIdRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetProfileProfileIdComponentComponentId")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(GetProfileProfileIdComponentComponentIdResponseObject); ok {
		if err := validResponse.VisitGetProfileProfileIdComponentComponentIdResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// PatchProfileProfileIdComponentComponentId operation middleware
func (sh *strictHandler) PatchProfileProfileIdComponentComponentId(w http.ResponseWriter, r *http.Request, profileId string, componentId string) {
	var request PatchProfileProfileIdComponentComponentIdRequestObject

	request.ProfileId = profileId
	request.ComponentId = componentId

	var body PatchProfileProfileIdComponentComponentIdJSONRequestBody
	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
		sh.options.RequestErrorHandlerFunc(w, r, fmt.Errorf("can't decode JSON body: %w", err))
		return
	}
	request.Body = &body

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.PatchProfileProfileIdComponentComponentId(ctx, request.(PatchProfileProfileIdComponentComponentIdRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "PatchProfileProfileIdComponentComponentId")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(PatchProfileProfileIdComponentComponentIdResponseObject); ok {
		if err := validResponse.VisitPatchProfileProfileIdComponentComponentIdResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// PostProfileProfileIdRunRunId operation middleware
func (sh *strictHandler) PostProfileProfileIdRunRunId(w http.ResponseWriter, r *http.Request, profileId string, runId string) {
	var request PostProfileProfileIdRunRunIdRequestObject

	request.ProfileId = profileId
	request.RunId = runId

	var body PostProfileProfileIdRunRunIdJSONRequestBody
	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
		sh.options.RequestErrorHandlerFunc(w, r, fmt.Errorf("can't decode JSON body: %w", err))
		return
	}
	request.Body = &body

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.PostProfileProfileIdRunRunId(ctx, request.(PostProfileProfileIdRunRunIdRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "PostProfileProfileIdRunRunId")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(PostProfileProfileIdRunRunIdResponseObject); ok {
		if err := validResponse.VisitPostProfileProfileIdRunRunIdResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}